split and move
